# src/api/api.yaml
openapi: 3.0.0
info:
  title: Senior Design Project API
  description: API for managing beam data, machines, results, baselines, and thresholds
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
...
paths:
...
  /machines/{machineId}:
    put:
      tags:
        - Machines
      summary: Update a machine
      description: Updates machine information (e.g., location)
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
          description: Machine identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Machine'
      responses:
        '200':
          description: Machine updated successfully
        '400':
          description: Bad request
        '404':
          description: Machine not found
    
    delete:
      tags:
        - Machines
      summary: Delete a machine
      description: Removes a machine from the system
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
          description: Machine identifier
      responses:
        '204':
          description: Machine deleted successfully
        '404':
          description: Machine not found
...
components:
  schemas:
...
    Machine:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the machine
        location:
          type: string
          description: Physical location of the machine
        name:
          type: string
          description: Name of the machine
        type:
          type: string
          description: Type of machine
      required:
        - id
        - location
        - name
        - type
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []